requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==2.2.13
django-debug-toolbar==2.2
django-environ==0.4.5
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
python-dotenv==0.12.0
pytz==2019.3
requests==2.22.0
sentry-sdk==0.16.2
six==1.13.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
zipp==2.2.0

twilio~=6.58.0